+++ ModelSwapperSkins/BoneMapping/InitializerRules/BoneInitializerRules_Loader.cs
@@ -1,0 +1,36 @@
+﻿using RoR2;
+using UnityEngine;
+
+namespace ModelSwapperSkins.BoneMapping.InitializerRules
+{
+    public class BoneInitializerRules_Loader : BoneInitializerRules_AutoName
+    {
+        public static new BoneInitializerRules_Loader Instance { get; } = new BoneInitializerRules_Loader();
+
+        BoneInitializerRules_Loader()
+        {
+        }
+
+        public override bool AppliesTo(BodyIndex bodyIndex)
+        {
+            return bodyIndex == BodyCatalog.FindBodyIndex("LoaderBody");
+        }
+
+        protected override BoneInfo getBoneInfo(Transform modelTransform, Transform potentialBoneTransform)
+        {
+            BoneInfo bone = base.getBoneInfo(modelTransform, potentialBoneTransform);
+
+            switch (bone.Type)
+            {
+                case BoneType.ShoulderL:
+                    bone.RotationOffset *= Quaternion.Euler(0f, 90f, 0f);
+                    break;
+                case BoneType.ShoulderR:
+                    bone.RotationOffset *= Quaternion.Euler(0f, 270f, 0f);
+                    break;
+            }
+
+            return bone;
+        }
+    }
+}